+++ base/src/main/java/com/qsmaxmin/qsbase/mvp/fragment/QsPullRecyclerFragment.java
@@ -25,8 +25,8 @@ import java.util.List;
  */
 
 public abstract class QsPullRecyclerFragment<P extends QsPresenter, D> extends QsRecyclerFragment<P, D> implements QsIPullToRefresh {
-    public static final int     LOAD_WHEN_SCROLL_TO_BOTTOM = 0;
-    public static final int     LOAD_WHEN_SECOND_TO_LAST   = 1;
+    public static final byte    LOAD_WHEN_SCROLL_TO_BOTTOM = 0;
+    public static final byte    LOAD_WHEN_SECOND_TO_LAST   = 1;
     private             boolean canLoadingMore             = true;
     private   PtrFrameLayout mPtrFrameLayout;
     protected LoadingFooter  mLoadingFooter;
